Implement Windows NT `LIST_ENTRY` in idiomatic Rust, safe where possible
I've added traits to uniquely identify lists. This lets Rust's type system take care of modifying the correct `ListEntry` field of an element structure.
A `BoxingListHead` enables entirely safe use of Windows NT doubly-linked lists by boxing each element on insertion and thereby establishing clear ownership of the elements.
The implementation also comes with constant and mutable iterators that implement `DoubleEndedIterator` and `FusedIterator`.
I use pinning to ensure that the list header address doesn't change, because the end marker of a list is always a reference back to its header.
And finally, I have also considered sharing an element in multiple linked lists, something that is sometimes required in Windows.