Add `NtSingleListHead`, `NtBoxingSingleListHead` for `SINGLE_LIST_ENTRY`
A Rust implementation of `SINGLE_LIST_ENTRY` is much simpler:
It never references back to itself, so pinning is not necessary.
With only a single link in forward direction, we also cannot provide as many features as for the doubly-linked list (`LIST_ENTRY`).
This means pushing and popping only from the front, no `DoubleEndedIterator`, and no `append`.