Implement `Extend` for `NtBoxingListHead`.
Unlike `NtBoxingSingleListHead`, we cannot feasibly implement `FromIterator` here, because the list head is pinned and shall usually reside on the stack.
`FromIterator` would create a new list that is inevitably moved on return and thereby invalid.

However, a doubly-linked list knows about the last element and can jump to it in O(1).
Hence, we can implement the even more versatile `Extend` trait to push a series of elements to an empty or an existing list.