Derive mutable pointers from mutable references
Deriving a mutable pointer from an immutable reference results in a `Frozen`
mutable pointer, which is UB to write through. Since all of the contexts
that a mutable pointer is needed also have a mutable reference available,
we can derive the pointer from the mutable reference instead.

This fixes TB violations.