Fix out-of-range accesses from getting the containing record
The issue with the original `containing_record{,_mut}` methods under SB is that
the pointer derived from the list entry reference is only valid for the range
of list entry, and not the containing record itself. By deferring the raw
pointer dereference to once we have a pointer to the containing record, the
resultant reference is valid for the whole record.

This fixes all violations under SB.